+++ src/main/webapp/quickstart/src/app/app-routing.module.ts
@@ -2,14 +2,14 @@ import { NgModule }             from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { DashboardComponent }   from './dashboard.component';
-import { HeroesComponent }      from './contractors.component';
+import { ContractorsComponent }      from './contractors.component';
 import { HeroDetailComponent }  from './hero-detail.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/dashboard', pathMatch: 'full' },
   { path: 'dashboard',  component: DashboardComponent },
   { path: 'detail/:id', component: HeroDetailComponent },
-  { path: 'contractors',     component: HeroesComponent }
+  { path: 'contractors',     component: ContractorsComponent }
 ];
 
 @NgModule({