+++ src/main/webapp/quickstart/CHANGELOG.md
@@ -45,7 +45,7 @@ Upgraders: for a fresh start, consider running these commands
   * removed wallaby
   * shrink styles.css
   * refine tsconfig.json
-  * `HeroesComponent` uses interpolation
+  * `ContractorsComponent` uses interpolation
 
 <a name="0.2.16"></a>
 # 0.2.16 (2016-11-14)